+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog for WindowsUpdateSetting
 
+## v0.3.0
+
++ Updated ShouldProcess message on `Suspend-WindowsUpdate` (Issue #3)
++ Added custom type names to PSCustomobject output (Issue #2)
++ Updated README
++ Updated documentation
+
 ## v0.2.1
 
 + Github updates